More code smells and light refactoring
