+++ src/main/java/com/revature/AfLocationServiceApplication.java
@@ -1,4 +1,4 @@
-package com.revature.AFLocationService;
+package com.revature;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
