+++ src/main/java/com/revature/dtos/UserDto.java
@@ -11,6 +11,7 @@ public class UserDto {
     private String role;
 
     public UserDto() {
+        //empty constructor for java bean implementation
     }
 
     public int getId() {
