+++ src/main/resources/application-test.properties
@@ -1,0 +1,8 @@
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.datasource.initialize=true
+spring.h2.console.enabled=true
+spring.jpa.hibernate.ddl-auto=none
+server.error.include-message=always
