+++ src/main/resources/application.properties
@@ -1,11 +1,7 @@
-spring.datasource.url=jdbc:h2:mem:testdb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
+spring.datasource.url=jdbc:postgresql://34.66.74.89:5432/location-reservation-db
+spring.datasource.driverClassName=org.postgresql.Driver
+spring.datasource.username=revatureBatch944
+spring.datasource.password=AdamRanieri
 spring.datasource.initialize=true
-spring.h2.console.enabled=true
 spring.jpa.hibernate.ddl-auto=none
 server.error.include-message=always
-
-
-
