+++ src/test/java/com/revature/controllers/LocationControllerTest.java
@@ -1,6 +1,6 @@
 package com.revature.controllers;
 
-import com.revature.AFLocationService.AfLocationServiceApplication;
+import com.revature.AfLocationServiceApplication;
 import com.revature.entities.Location;
 import com.revature.exceptions.LocationNotFoundException;
 import com.revature.services.LocationService;
