+++ src/test/java/com/revature/controllers/RoomControllerTests.java
@@ -1,7 +1,7 @@
 package com.revature.controllers;
 
 
-import com.revature.AFLocationService.AfLocationServiceApplication;
+import com.revature.AfLocationServiceApplication;
 import com.revature.entities.Room;
 import com.revature.entities.RoomType;
 import com.revature.exceptions.RoomNotFoundException;
