+++ src/test/java/com/revature/repos/RoomRepoTests.java
@@ -1,16 +1,18 @@
 package com.revature.repos;
 
-import com.revature.AFLocationService.AfLocationServiceApplication;
+import com.revature.AfLocationServiceApplication;
 import com.revature.entities.Room;
 import com.revature.entities.RoomType;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 import java.util.ArrayList;
 import java.util.List;
 @SpringBootTest(classes= AfLocationServiceApplication.class)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+@ActiveProfiles("test")
 class RoomRepoTests {
 
     @Autowired
