+++ src/test/java/com/revature/services/RoomServiceTests.java
@@ -1,6 +1,6 @@
 package com.revature.services;
 
-import com.revature.AFLocationService.AfLocationServiceApplication;
+import com.revature.AfLocationServiceApplication;
 import com.revature.entities.Building;
 import com.revature.entities.Location;
 import com.revature.entities.Room;
@@ -15,6 +15,7 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.ActiveProfiles;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,6 +25,7 @@ import java.util.Optional;
 @ExtendWith(MockitoExtension.class)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@ActiveProfiles("test")
 class RoomServiceTests {
 
     @InjectMocks
